+++ src/cheat/precompiled.h
@@ -284,6 +284,7 @@ namespace hl
 #include "features/AntiDemoRecorder.h"
 #include "features/WorldVisuals.h"
 #include "features/AntiScreen.h"
+#include "features/AspectRatioChanger.h"
 #include "features/movement/BunnyHop.h"
 #include "features/movement/AirStuck.h"
 #include "features/movement/GroundStrafe.h"