+++ changelog.md
@@ -146,6 +146,7 @@ Sorry for the wait, I've been busy with IRL stuff. 🙄😪
 ## Additions and improvements
 
 - Implemented Anti-Screen.
+- Implemented aspect ratio changer.
 
 ## Bugfixes